website/cy/requirements.txt
Django==1.9.13

# db related
psycopg2==2.6.2

# api related
djangorestframework==3.3.1
django-filter==0.11

# deploy related
gunicorn==19.6.0

# testing related
coverage==3.7.1

# elastic search
elasticsearch==1.6.0
django-haystack==2.4.0

# others
django-pagination==1.0.7
django-debug-toolbar==1.5
